products fields changed to float

+++ base/models.py
@@ -13,8 +13,8 @@ class Product(models.Model):
     name = models.CharField(max_length=255)
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     stock = models.IntegerField()
-    purchase_price = models.DecimalField(max_digits=10, decimal_places=2)
-    selling_price = models.DecimalField(max_digits=10, decimal_places=2)
+    purchase_price = models.FloatField()
+    selling_price = models.FloatField()
 
     def __str__(self):
         return self.name
